OJ-1960: Set SonarCloud base dir by default to avoid permissions issue
The SonarCloud cleanup script needs write permissions but the default base dir is the root of the file system where it cannot run the "chown" command. Set the base dir to the location where the repository is checked out by default to prevent the job from failing due to lack of permissions.