BAU: Reset SAM cache when the dependencies change
Cache SAM dependencies based on the template file, as the cache should be refreshed if the template file has changed.

Allow to specify a source code directory so the cache can be refreshed when the sources change.

Return the cache keys generated in the SAM build action so they can be passed to the deploy action.