SPT-1521: Do not error if no matching stacks
Update the `grep` expression for matching `$name_filter` so that the
command succeeds even if there are no matches, the rest of the script
copes correctly if there is nothing to do.

Teams using this action are seeing jobs failing when there is nothing
to clean up, it would be neater for the job to succeed and do nothing.